ALL API's Live from AI Services- (abhi)

+++ Backend2/Modules/Document/Controllers/deleteDocumentByIdController.js
@@ -1,5 +1,6 @@
 import Document from "../Mongo_Models/DocUploadModel.js";
 import client from "../../../config/sqlDB.js";
+import axios from "axios";
 import { getDocument } from "../Models/getDocumentTableModel.js";
 import { deleteMetadata } from "../Models/deleteMetadataModel.js";
 import { deleteEntity } from "../Models/deleteEntityModel.js";
@@ -21,21 +22,35 @@ export const deleteDocumentByIdController = async (req, res) => {
     // delete from document
     const documentData = await deleteDocument(document_id);
     // save temp data from mongo db and delete the mongo data
-    MongoData = await Document.findByIdAndDelete(Doc.mongodb_uid);
 
     if (
       metaData.success &&
       entityData.success &&
       documentOwnerMap.success &&
-      MongoData &&
       documentData.success
     ) {
-      client.query("COMMIT");
-      res.status(200).json({
-        success: true,
-        message: "Document deleted successfully",
-        data: Doc,
+      // document delete api from python
+      // http://13.127.253.239:5004/delete
+
+      const response = await axios.post(`http://${process.env.RAG}/delete`, {
+        doc_id: document_id,
       });
+
+      if (response.data.deleted) {
+        MongoData = await Document.findByIdAndDelete(Doc.mongodb_uid);
+        if (!MongoData) {
+          throw new Error("MongoDB Document Deletion Failed");
+        }
+        client.query("COMMIT");
+        res.status(200).json({
+          success: true,
+          message: "Document deleted successfully",
+          data: Doc,
+          data_vector: response.data,
+        });
+      } else {
+        throw new Error("Document Deletion from Vector Database Failed");
+      }
     } else {
       throw new Error("Document Deletion Failed");
     }
